requirements.txt
scikit-learn
numpy
alembic
fastapi
uvicorn
sqlalchemy
pydantic
PyJWT
bcrypt
passlib
asyncpg
pandas
psycopg2-binary
python-dotenv